Modularize database models
