+++ server/users/models.js
@@ -1,0 +1,13 @@
+module.exports = function(sequelize, DataTypes) {
+  return sequelize.define('User', {
+    id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
+    username: DataTypes.STRING,
+    firstname: DataTypes.STRING,
+    lastname: DataTypes.STRING,
+    reputation: DataTypes.INTEGER,
+    beebucks: DataTypes.INTEGER,
+    city: DataTypes.STRING,
+    state: DataTypes.STRING,
+    country: DataTypes.STRING
+  });
+}
